pyproject.toml
[project]
name = "spatialbench"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "duckdb>=1.2",
    "pydantic>=2.7",
    "requests>=2.32",
    "pyyaml>=6.0",
]

[dependency-groups]
dev = [
    "hypothesis>=6.100",
    "mypy>=1.10",
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.6",
    "types-PyYAML>=6.0",
    "types-requests>=2.32",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=spatialbench --cov-branch --cov-report=term-missing --cov-fail-under=100"

[tool.mypy]
strict = true
python_version = "3.12"
plugins = ["pydantic.mypy"]

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B", "SIM", "RUF"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
